chainstate-storage: Put test into a separate module
